+++ cloud/app.js
@@ -51,6 +51,9 @@ app.use('/', require('cloud/user'));
 // Settings
 app.use('/', require('cloud/settings'));
 
+// Resume
+app.use('/', require('cloud/resume'));
+
 // // Example reading from the request query string of an HTTP get request.
 // app.get('/test', function(req, res) {
 //   // GET http://example.parseapp.com/test?message=hello
