+++ cloud/views/settings.ejs
@@ -1,6 +1,6 @@
 <div class="row">
   <div class="col-lg-12">
-    <form role="form" method="post" action="/update-settings">
+    <form role="form" method="post" action="/update-settings" id="settings-form">
       <div class="form-group">
         <label for="firstName">First Name</label>
         <input type="text" class="form-control" id="firstName" name='firstName' placeholder="Your First Name" value="<%= firstName %>">
@@ -18,8 +18,17 @@
         <input type="url" class="form-control" id="linkedin" name='linkedin' placeholder="Enter Linked In URL" value="<%= linkedin %>">
       </div>
       <div class="form-group">
+        <label for="resumeUploadFile">Resume Upload</label>
+          <input type="file" name="resumeUploadFile" id="resumeUploadFile">
+          <p class="help-block">Upload your resume in PDF format</p>
+          <a href="#" class="upload">Upload</a>
+      </div>
+      <div class="form-group">
         <button type="submit" class="btn btn-default">Submit</button>
       </div>
     </form>
+    <script type="text/javascript">
+      new Uploader({ el: "#settings-form" });
+    </script>
   </div>
 </div>
