+++ public/javascripts/main.js
@@ -4,20 +4,23 @@ Uploader = Backbone.View.extend({
     "change input[type=file]": "upload",
     "click .upload": "showFile"
   },
-  
+
   initialize: function() {
+    console.log('init');
     var self = this;
     this.fileUploadControl = this.$el.find("input[type=file]")[0];
   },
 
   showFile: function(e) {
+    console.log('show file');
     this.fileUploadControl.click();
     return false;
   },
 
   upload: function() {
+    console.log('upload');
     var self = this;
-    
+
     if (this.fileUploadControl.files.length > 0) {
       this.$(".upload").html("Uploading <img src='/images/spinner.gif' />");
       var file = this.fileUploadControl.files[0];
@@ -28,7 +31,7 @@ Uploader = Backbone.View.extend({
       parseFile.save().then(function() {
         // Then, we post to our custom endpoint which will do the post
         // processing necessary for the image page
-        $.post("/i", {
+        $.post("/upload", {
           file: {
             "__type": "File",
             "url": parseFile.url(),
@@ -53,9 +56,11 @@ Uploader = Backbone.View.extend({
 $(function() {
   // Make all of special links magically post the form
   // when it has a particular data-action associated
-  $("a[data-action='post']").click(function(e) {
+  $("button[type='submit']").click(function(e) {
     var el = $(e.target);
-    el.closest("form").submit();
+    console.log($(e.target));
+    //el.closest("form").submit();
+    //return false;
     return false;
   });
 });
